Add support for Django 3.2
