+++ cacheops/__init__.py
@@ -1,21 +1,11 @@
 __version__ = '5.0.1'
 VERSION = tuple(map(int, __version__.split('.')))
 
-
-from django.apps import AppConfig
-
+import django
 from .simple import *
 from .query import *
 from .invalidation import *
 from .templatetags.cacheops import *
-from .transaction import install_cacheops_transaction_support
 
-
-class CacheopsConfig(AppConfig):
-    name = 'cacheops'
-
-    def ready(self):
-        install_cacheops()
-        install_cacheops_transaction_support()
-
-default_app_config = 'cacheops.CacheopsConfig'
+if django.VERSION < (3, 2):
+    default_app_config = "cacheops.apps.CacheopsConfig"
